requirements.txt
# SociClaw Dependencies
requests>=2.31.0
py-trello>=0.19.0
notion-client>=2.2.0
python-dotenv>=1.0.0
tweepy>=4.14.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
